renamed ros node to vp_engine_ros to avoid conflict with python package vp_engine
