+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>vp_engine</name>
+  <name>vp_engine_ros</name>
   <version>0.1.0</version>
   <description>Visual Perception Engine</description>
   <maintainer email="rob.royce@jpl.nasa.gov">Rob Royce</maintainer>
